backend/pyproject.toml
[project]
name = "languine"
version = "0.1.0"
description = "Backend for Languine App"
authors = [
    {name = "Marwa Khafagy"}, 
    {name = "Selina Kim"}, 
    {name = "Giovanna Gerada"}, 
    {name = "Stella Gu"}, 
    {name = "Tinu Joseph"}
]
license = "MIT"
readme = "README.md"
requires-python = ">=3.11.0,<3.14"
dynamic = ["dependencies"]

[tool.poetry]

[tool.poetry.dependencies]
python = ">=3.11.0,<3.14"
flask = "^3.1.2"
pydantic = "^2.12.2"
flask-jwt-extended = "^4.7.1"
flask-limiter = "^4.0.0"
flask-sqlalchemy = "^3.1.1"
requests = "^2.32.5"
deepl = "^1.25.0"
python-dotenv = "^1.2.1"
pandas = "^3.0.1"
tqdm = "^4.67.3"
coqui-tts = "^0.27.3"
scipy = "^1.11.0"
cutlet = "^0.5.0"
flask-cors = "^6.0.1"
google-auth-oauthlib = "^1.2.3"
google-auth-httplib2 = "^0.2.1"
psycopg2-binary = "^2.9.11"
pypinyin = "^0.53.0"
unidic-lite = "^1.0.8"


[tool.poetry.group.dev.dependencies]
pytest = "^8.4.2"
pytest-flask = "^1.3.0"
pytest-cov = "^7.0.0"

[tool.pytest.ini_options]
pythonpath = ["src"]
addopts = "-m 'not integration'"
markers = [
    "integration: marks tests as integration tests (deselect with '-m \"not integration\"')"
]

[build-system]
requires = ["poetry-core>=2.2.0"]
build-backend = "poetry.core.masonry.api"
